requirements.txt
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.0
daiquiri==3.2.1
docopt==0.6.2
docutils==0.19
fernet==1.0.1
idna==3.3
Kivy==2.1.0
kivy-deps.angle==0.3.2
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
mysql==0.0.3
mysql-connector-python==8.0.30
mysqlclient==2.1.1
Pillow==9.2.0
protobuf==3.20.1
pyaes==1.6.1
pycodestyle==2.9.1
Pygments==2.13.0
pypiwin32==223
python-dotenv==0.20.0
python-json-logger==2.0.4
pywin32==304
requests==2.28.1
toml==0.10.2
urllib3==1.26.11
yarg==0.1.9
